requirements.txt
streamlit==1.22.0
faiss-cpu==1.7.4
sentence-transformers==2.2.2
pyPDF2
python-dotenv
langchain_groq
lanchain_community
python-dotenv==1.0.0
pydantic==1.10.7
tiktoken==0.4.0
langchain_core
langchain
